Removed unused or irrelevant specs
Because this is only intended to be used as a design prototyping kit, we do not believe there is a requirement for specs to cover the code base. We have left rspec in for now so the option is there to include specs for anything you develop, but unlike a production system it is not seen as critical by the originating team.