Remove devise from the kit and update code to work without it
We're switching to a simpler model for the prototype kit which does not rely on interacting with the database, or the use of configuring something like devise. If users wish to mock an authentication system (including devise implies they will which may not actually be the case) they should steer to using the dummy_data functionality and their own custom views.