Add ability to simply restrict access to the site to only permitted users
The idea is that you are probably using a service like heroku to host your prototypes to enable users to access it in order to give feedback. However either because you don't believe the prototype is ready, or because it exposes sensitive functionality or content you want to be able to restrict access to it. We have intentionally created a simple way to do this, that just requires creating a single shared key. Anyone you give the key to will then be able to access the site, everyone else will get the public/private.html.erb page which you can edit to your liking. More details will be available in the GitHub repo.