Replace call to env variable for production secret_key_base to fixed value
The kit is not intended to be used as a production system, with the priority being to make it as simply to use as possible. Therefore update the prod secret_key_base to be a set value so that should a user attempt to run it in production mode they won't get an error because they didn't know to also create the relevant environment variable.