Add vagrant file to enable easier environment setup for new users
To enable new users and those unfamilier to rails projects to get up and running quickly, we have enabled Vagrant within the project. Vagrant is a tool that enables the provisioning of virtual machines from a config file. This means now when a user grabs the project, as long as they have vagrant installed they will be able to use it to provision the right environment, rather than figure out how to set it up for themselves. We also plan to expand the README and github wiki with supplementary information.