Correct rails version in README
Having updated the version of Rails in a previous PR (#46) we forgot to update the version of rails referenced in the README.

This change corrects that, as well as choosing to not be so specific. We now feel its not necessary to explicitly state the reference in the README, and should lead to less errors keeping the documentation up to date.