Add proto_controller show test
Whilst the project is not being actively worked on we use Deppbot to ensure our dependencies are up to date.

However recently when returning to the project we found the static pages were no longer loading. Therefore this change adds a test to check that all static pages return 'ok' when called so our CI server can hopefully flag an issue, rather than waiting to find it for ourselves.