Update ref. factory_girl to factory_bot
[ThoughtBot](https://thoughtbot.com/) the maintainers of the **factory_girl** gem recently decided to rename it to **factory_bot**. However as this was done on a fix version (0.0.x) its meant that a number of our projects see this as a potential update.

To get it to work though you not only need to change the reference in the Gemfile, but also how you reference it in the code. Hence any automated dependency checking services are suggesting the project takes the update but then CI is failing.

This change manually updates our reference to factory girl to use [factory_bot](https://github.com/thoughtbot/factory_bot) instead, and updates any references in the code to get the tests passing again.